Langgraph_Question_Answer_System-main/pyproject.toml
[project]
name = "langgraph-qa-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
